requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.1.7
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==4.11.3
python-dotenv==1.0.1
pytz==2025.1
setuptools==77.0.3
sqlparse==0.5.3
